python/requirements.txt
numpy
pandas
transformers
torch
torchvision
accelerate
pulsar-client